requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.2
django-autoslug-iplweb==1.9.4
django-cors-headers==2.4.0
django-extensions==2.1.3
django-smtp-ssl==1.0
django-versatileimagefield==1.10
djangorestframework==3.9.0
factory-boy==2.11.1
Faker==0.9.2
Flask==1.0.2
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
sendgrid==5.6.0
six==1.11.0
social-auth-app-django==3.0.0
social-auth-core==2.0.0
text-unidecode==1.2
urllib3==1.24
virtualenv==16.0.0
Werkzeug==0.14.1
whitenoise==3.2.1
